+++ isChromatic.js
@@ -1,2 +1,8 @@
-/* eslint-disable */
-module.exports = require('./dist/isChromatic');
+/* eslint-env browser */
+
+module.exports = function isChromatic() {
+  return !!(
+    typeof window !== 'undefined' &&
+    (window.navigator.userAgent.match(/Chromatic/) || window.location.href.match(/chromatic=true/))
+  );
+};